refactor(inputs): use onInfoClick props + remove Wrapper from component name
