+++ libs/ui/src/components/RecordEdition/EditRecordContent/uiElements/StandardField/StandardFieldValue/DSInput.test.tsx
@@ -2,7 +2,7 @@
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
 import {render, screen} from '_ui/_tests/testUtils';
-import {DSInputPasswordWrapper} from './DSInputPasswordWrapper';
+import {DSInput} from './DSInput';
 import {FieldScope} from '../../../_types';
 import {
     InheritedFlags,
@@ -15,7 +15,6 @@ import {mockAttributeLink} from '_ui/__mocks__/common/attribute';
 import userEvent from '@testing-library/user-event';
 import {AntForm} from 'aristid-ds';
 import {RecordFormAttributeFragment} from '_ui/_gqlTypes';
-import {RecordFormElementsValueStandardValue} from '_ui/hooks/useGetRecordForm';
 
 const en_label = 'label';
 const fr_label = 'libellé';
@@ -23,7 +22,7 @@ const idValue = '123';
 const mockValue = {
     index: 0,
     displayValue: 'my value',
-    editingValue: 'my editing value',
+    editingValue: 'my raw value',
     originRawValue: 'my raw value',
     idValue: null,
     isEditing: false,
@@ -42,13 +41,7 @@ const mockValue = {
     state: StandardFieldValueState.PRISTINE
 };
 
-const getInitialState = ({
-    required,
-    fallbackLang
-}: {
-    required: boolean;
-    fallbackLang: boolean;
-}): IStandardFieldReducerState => ({
+const getInitialState = (required: boolean, fallbackLang = false): IStandardFieldReducerState => ({
     record: mockRecord,
     formElement: {
         ...mockFormElementInput,
@@ -74,7 +67,7 @@ const getInitialState = ({
     isInheritedValue: false
 });
 
-const inheritedValues: RecordFormElementsValueStandardValue[] = [
+const inheritedValues = [
     {
         isInherited: null,
         value: 'override value',
@@ -101,7 +94,7 @@ const inheritedOverrideValue: InheritedFlags = {
     inheritedValue: {raw_value: inheritedValues[1].raw_value}
 };
 
-describe('DSInputPasswordWrapper', () => {
+describe('DSInput', () => {
     const mockHandleSubmit = jest.fn();
     const mockOnChange = jest.fn();
     let user!: ReturnType<typeof userEvent.setup>;
@@ -113,11 +106,11 @@ describe('DSInputPasswordWrapper', () => {
     });
 
     test('Should display input with fr label ', async () => {
-        const state = getInitialState({required: false, fallbackLang: false});
+        const state = getInitialState(false, false);
         render(
             <AntForm>
                 <AntForm.Item>
-                    <DSInputPasswordWrapper
+                    <DSInput
                         state={state}
                         attribute={{} as RecordFormAttributeFragment}
                         fieldValue={null}
@@ -132,11 +125,11 @@ describe('DSInputPasswordWrapper', () => {
     });
 
     test('Should display input with fallback label ', async () => {
-        const state = getInitialState({required: false, fallbackLang: true});
+        const state = getInitialState(false, true);
         render(
             <AntForm>
                 <AntForm.Item>
-                    <DSInputPasswordWrapper
+                    <DSInput
                         state={state}
                         attribute={{} as RecordFormAttributeFragment}
                         fieldValue={null}
@@ -151,11 +144,11 @@ describe('DSInputPasswordWrapper', () => {
     });
 
     test('Should submit empty value if field is not required', async () => {
-        const state = getInitialState({required: false, fallbackLang: false});
+        const state = getInitialState(false);
         render(
             <AntForm>
                 <AntForm.Item>
-                    <DSInputPasswordWrapper
+                    <DSInput
                         state={state}
                         attribute={{} as RecordFormAttributeFragment}
                         fieldValue={null}
@@ -166,22 +159,21 @@ describe('DSInputPasswordWrapper', () => {
             </AntForm>
         );
 
-        // TODO : change testid with click on label when https://aristid.atlassian.net/browse/DS-174 is done
-        const input = screen.getByTestId('kit-input-password');
+        const input = screen.getByRole('textbox');
         await user.click(input);
         await user.tab();
 
         expect(mockHandleSubmit).toHaveBeenCalledWith('', state.attribute.id);
-        expect(mockOnChange).toHaveBeenCalledTimes(1);
+        expect(mockOnChange).toHaveBeenCalled();
     });
 
     describe('With required input and no inheritance', () => {
         test('Should submit the value if field is not empty', async () => {
-            const state = getInitialState({required: true, fallbackLang: false});
+            const state = getInitialState(true);
             render(
                 <AntForm>
                     <AntForm.Item>
-                        <DSInputPasswordWrapper
+                        <DSInput
                             state={state}
                             attribute={{} as RecordFormAttributeFragment}
                             fieldValue={null}
@@ -193,7 +185,7 @@ describe('DSInputPasswordWrapper', () => {
             );
 
             const text = 'text';
-            const input = screen.getByTestId('kit-input-password');
+            const input = screen.getByRole('textbox');
             await user.click(input);
             await user.type(input, text);
             await user.tab();
@@ -203,11 +195,11 @@ describe('DSInputPasswordWrapper', () => {
         });
 
         test('Should submit the default value if field is empty', async () => {
-            const state = getInitialState({required: true, fallbackLang: false});
+            const state = getInitialState(true);
             render(
                 <AntForm>
                     <AntForm.Item>
-                        <DSInputPasswordWrapper
+                        <DSInput
                             state={state}
                             attribute={{} as RecordFormAttributeFragment}
                             fieldValue={null}
@@ -219,7 +211,7 @@ describe('DSInputPasswordWrapper', () => {
                 </AntForm>
             );
 
-            const input = screen.getByTestId('kit-input-password');
+            const input = screen.getByRole('textbox');
             await user.click(input);
             await user.tab();
 
@@ -229,7 +221,7 @@ describe('DSInputPasswordWrapper', () => {
 
     describe('With inheritance', () => {
         test("Should display the inherited value by default and not save if we don't change it", async () => {
-            let state = getInitialState({required: false, fallbackLang: false});
+            let state = getInitialState(false);
             state = {
                 ...state,
                 ...inheritedNotOverrideValue,
@@ -238,7 +230,7 @@ describe('DSInputPasswordWrapper', () => {
             render(
                 <AntForm>
                     <AntForm.Item>
-                        <DSInputPasswordWrapper
+                        <DSInput
                             state={state}
                             attribute={{} as RecordFormAttributeFragment}
                             fieldValue={null}
@@ -249,7 +241,7 @@ describe('DSInputPasswordWrapper', () => {
                     </AntForm.Item>
                 </AntForm>
             );
-            const input = screen.getByTestId('kit-input-password');
+            const input = screen.getByRole('textbox');
             expect(input).toHaveValue(inheritedValues[1].raw_value);
 
             expect(screen.queryByRole('button')).toBeNull();
@@ -261,7 +253,7 @@ describe('DSInputPasswordWrapper', () => {
         });
 
         test('Should display the override value in the input and inherited value under it', async () => {
-            let state = getInitialState({required: false, fallbackLang: false});
+            let state = getInitialState(false);
             state = {
                 ...state,
                 ...inheritedOverrideValue,
@@ -271,7 +263,7 @@ describe('DSInputPasswordWrapper', () => {
             render(
                 <AntForm>
                     <AntForm.Item>
-                        <DSInputPasswordWrapper
+                        <DSInput
                             state={state}
                             attribute={{} as RecordFormAttributeFragment}
                             fieldValue={null}
@@ -283,14 +275,14 @@ describe('DSInputPasswordWrapper', () => {
                 </AntForm>
             );
 
-            const input = screen.getByTestId('kit-input-password');
+            const input = screen.getByRole('textbox');
             const helperText = screen.getByText(/inherited value/);
             expect(input).toHaveValue(inheritedValues[0].raw_value);
             expect(helperText).toBeInTheDocument();
         });
 
         test("Should allow to clear input when it's override", async () => {
-            let state = getInitialState({required: false, fallbackLang: false});
+            let state = getInitialState(false);
             state = {
                 ...state,
                 ...inheritedOverrideValue,
@@ -299,7 +291,7 @@ describe('DSInputPasswordWrapper', () => {
             render(
                 <AntForm>
                     <AntForm.Item>
-                        <DSInputPasswordWrapper
+                        <DSInput
                             state={state}
                             attribute={{} as RecordFormAttributeFragment}
                             fieldValue={null}
