+++ libs/ui/src/components/RecordEdition/EditRecordContent/uiElements/StandardField/StandardFieldValue/DSInput.tsx
@@ -1,58 +1,48 @@
 // Copyright LEAV Solutions 2017
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {KitInputNumber, KitInputWrapper} from 'aristid-ds';
-import {ComponentPropsWithRef, FocusEvent, FunctionComponent, useState} from 'react';
+import {KitInput} from 'aristid-ds';
+import {ChangeEvent, FocusEvent, FunctionComponent, ReactNode, useState} from 'react';
 import {
     IStandardFieldReducerState,
     IStandardFieldValue
 } from '../../../reducers/standardFieldReducer/standardFieldReducer';
-import {Form, InputNumberProps} from 'antd';
+import {Form, InputProps} from 'antd';
 import {IProvidedByAntFormItem} from '_ui/components/RecordEdition/EditRecordContent/_types';
-import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
 import styled from 'styled-components';
+import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
 import {useValueDetailsButton} from '_ui/components/RecordEdition/EditRecordContent/shared/ValueDetailsBtn/useValueDetailsButton';
 import {RecordFormAttributeFragment} from '_ui/_gqlTypes';
-import {localizedTranslation} from '@leav/utils';
 import {useLang} from '_ui/hooks';
+import {localizedTranslation} from '@leav/utils';
 
-interface IDSInputWrapperProps extends IProvidedByAntFormItem<InputNumberProps> {
+interface IDSInputProps extends IProvidedByAntFormItem<InputProps> {
     state: IStandardFieldReducerState;
     attribute: RecordFormAttributeFragment;
     fieldValue: IStandardFieldValue;
-    shouldShowValueDetailsButton?: boolean;
     handleSubmit: (value: string, id?: string) => void;
 }
 
-const KitInputNumberStyled = styled(KitInputNumber)<{$shouldHighlightColor: boolean}>`
-    .ant-input-number-input-wrap .ant-input-number-input {
-        color: ${({$shouldHighlightColor}) =>
-            $shouldHighlightColor ? 'var(--general-colors-primary-400)' : 'initial'};
-    }
+const KitInputStyled = styled(KitInput)<{$shouldHighlightColor: boolean}>`
+    color: ${({$shouldHighlightColor}) => ($shouldHighlightColor ? 'var(--general-colors-primary-400)' : 'initial')};
 `;
 
-export const DSInputNumberWrapper: FunctionComponent<IDSInputWrapperProps> = ({
+export const DSInput: FunctionComponent<IDSInputProps> = ({
     value,
     onChange,
     state,
     attribute,
     fieldValue,
-    shouldShowValueDetailsButton = false,
     handleSubmit
 }) => {
-    if (!onChange) {
-        throw Error('DSInputNumberWrapper should be used inside a antd Form.Item');
-    }
-
     const {t} = useSharedTranslation();
-    const {lang} = useLang();
     const {errors} = Form.Item.useStatus();
     const {onValueDetailsButtonClick} = useValueDetailsButton({
         value: fieldValue?.value,
         attribute
     });
-
     const [hasChanged, setHasChanged] = useState(false);
+    const {lang: availableLang} = useLang();
 
     const _resetToInheritedValue = () => {
         setHasChanged(false);
@@ -69,18 +59,23 @@ export const DSInputNumberWrapper: FunctionComponent<IDSInputWrapperProps> = ({
         if (hasChanged || !state.isInheritedValue) {
             handleSubmit(valueToSubmit, state.attribute.id);
         }
-        onChange(valueToSubmit);
+        onChange(event);
     };
 
-    const _handleOnChange: ComponentPropsWithRef<typeof KitInputNumberStyled>['onChange'] = inputValue => {
+    const _handleOnChange = (event: ChangeEvent<HTMLInputElement>) => {
         setHasChanged(true);
-        onChange(inputValue);
+        const inputValue = event.target.value;
+        if (state.isInheritedValue && inputValue === '' && event.type === 'click') {
+            _resetToInheritedValue();
+            return;
+        }
+        onChange(event);
     };
 
-    const label = localizedTranslation(state.formElement.settings.label, lang);
+    const label = localizedTranslation(state.formElement.settings.label, availableLang);
 
     return (
-        <KitInputWrapper
+        <KitInputStyled
             required={state.formElement.settings.required}
             label={label}
             onInfoClick={onValueDetailsButtonClick}
@@ -92,14 +87,12 @@ export const DSInputNumberWrapper: FunctionComponent<IDSInputWrapperProps> = ({
                       })
                     : undefined
             }
-        >
-        <KitInputNumberStyled
             value={value}
-            onChange={_handleOnChange}
             disabled={state.isReadOnly}
+            allowClear={!state.isInheritedNotOverrideValue}
             onBlur={_handleOnBlur}
+            onChange={_handleOnChange}
             $shouldHighlightColor={!hasChanged && state.isInheritedNotOverrideValue}
         />
-        </KitInputWrapper>
     );
 };
